Correccion SoftDelete Implementada en Tabla

+++ src/main/java/com/jordi/hogwarts/Repository/CasaRepository.java
@@ -3,4 +3,4 @@ package com.jordi.hogwarts.Repository;
 import com.jordi.hogwarts.Model.Casa;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CasaRepository extends JpaRepository<Casa, Integer> {}
+public interface CasaRepository extends JpaRepository<Casa, Long> {}
