+++ src/main/java/com/jordi/hogwarts/Repository/MascotaRepository.java
@@ -3,4 +3,4 @@ package com.jordi.hogwarts.Repository;
 import com.jordi.hogwarts.Model.Mascota;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MascotaRepository extends JpaRepository<Mascota, Integer> {}
+public interface MascotaRepository extends JpaRepository<Mascota, Long> {}
